backend/requirements.txt
# Web framework
fastapi==0.109.1
uvicorn[standard]==0.23.2

# Database
sqlalchemy==2.0.22
psycopg2-binary==2.9.8  # PostgreSQL driver
alembic==1.12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.8.2

# Background tasks
celery[redis]==5.3.1
redis==5.3.7

# Cloud storage
boto3==1.28.41       # AWS S3
google-cloud-storage==2.12.0  # GCP
azure-storage-blob==12.23.0   # Azure

# Data validation
pydantic==2.3.0

# Audio processing
ffmpeg-python==0.2.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Optional: for type hints & development
mypy==1.5.1
types-redis==4.3.5
types-PyYAML==6.0.12
